sagemaker/model/robosat_pink/requirements.txt
torchvision
numpy
pillow
opencv-contrib-python
tqdm==4.19.9
requests
geojson
mercantile
osmium
rasterio
supermercado
shapely
rtree
pyproj
toml
pytest
webcolors
